backend/pyproject.toml
[project]
name = "todo-backend"
version = "0.1.0"
description = "Бекенд ToDo на Django и DRF"
requires-python = ">=3.11"
dependencies = [
    "Django==5.0",
    "djangorestframework",
    "psycopg2-binary",
    "celery",
    "redis",
]
